server/requirements.txt
flask_cors
flask
pandas
numpy
openpyxl
waitress
pymongo
pymongo[srv]
python-dotenv
requests
PyJWT
flask-jwt-extended